+++ model_api_key.go
@@ -9,10 +9,10 @@ type ApiKey struct {
 	Id          string    `json:"id"`
 	Secret      string    `json:"secret"`
 	Name        string    `json:"name"`
-	Nonce       float32   `json:"nonce"`
+	Nonce       int       `json:"nonce"`
 	Cidr        string    `json:"cidr,omitempty"`
 	Permissions []string  `json:"permissions,omitempty"`
 	Enabled     bool      `json:"enabled,omitempty"`
-	UserId      float32   `json:"userId"`
+	UserId      int       `json:"userId"`
 	Created     time.Time `json:"created,omitempty"`
 }
