+++ model_chat.go
@@ -6,11 +6,11 @@ import (
 
 // Trollbox Data
 type Chat struct {
-	Id        float32   `json:"id,omitempty"`
+	Id        int       `json:"id,omitempty"`
 	Date      time.Time `json:"date"`
 	User      string    `json:"user"`
 	Message   string    `json:"message"`
 	Html      string    `json:"html"`
 	FromBot   bool      `json:"fromBot,omitempty"`
-	ChannelID float64   `json:"channelID,omitempty"`
+	ChannelID int       `json:"channelID,omitempty"`
 }
