+++ model_execution.go
@@ -10,18 +10,18 @@ type Execution struct {
 	OrderID               string    `json:"orderID,omitempty"`
 	ClOrdID               string    `json:"clOrdID,omitempty"`
 	ClOrdLinkID           string    `json:"clOrdLinkID,omitempty"`
-	Account               float32   `json:"account,omitempty"`
+	Account               int       `json:"account,omitempty"`
 	Symbol                string    `json:"symbol,omitempty"`
 	Side                  string    `json:"side,omitempty"`
-	LastQty               float32   `json:"lastQty,omitempty"`
+	LastQty               int       `json:"lastQty,omitempty"`
 	LastPx                float64   `json:"lastPx,omitempty"`
 	UnderlyingLastPx      float64   `json:"underlyingLastPx,omitempty"`
 	LastMkt               string    `json:"lastMkt,omitempty"`
 	LastLiquidityInd      string    `json:"lastLiquidityInd,omitempty"`
 	SimpleOrderQty        float64   `json:"simpleOrderQty,omitempty"`
-	OrderQty              float32   `json:"orderQty,omitempty"`
+	OrderQty              int       `json:"orderQty,omitempty"`
 	Price                 float64   `json:"price,omitempty"`
-	DisplayQty            float32   `json:"displayQty,omitempty"`
+	DisplayQty            int       `json:"displayQty,omitempty"`
 	StopPx                float64   `json:"stopPx,omitempty"`
 	PegOffsetValue        float64   `json:"pegOffsetValue,omitempty"`
 	PegPriceType          string    `json:"pegPriceType,omitempty"`
@@ -38,17 +38,17 @@ type Execution struct {
 	WorkingIndicator      bool      `json:"workingIndicator,omitempty"`
 	OrdRejReason          string    `json:"ordRejReason,omitempty"`
 	SimpleLeavesQty       float64   `json:"simpleLeavesQty,omitempty"`
-	LeavesQty             float32   `json:"leavesQty,omitempty"`
+	LeavesQty             int       `json:"leavesQty,omitempty"`
 	SimpleCumQty          float64   `json:"simpleCumQty,omitempty"`
-	CumQty                float32   `json:"cumQty,omitempty"`
+	CumQty                int       `json:"cumQty,omitempty"`
 	AvgPx                 float64   `json:"avgPx,omitempty"`
 	Commission            float64   `json:"commission,omitempty"`
 	TradePublishIndicator string    `json:"tradePublishIndicator,omitempty"`
 	MultiLegReportingType string    `json:"multiLegReportingType,omitempty"`
 	Text                  string    `json:"text,omitempty"`
 	TrdMatchID            string    `json:"trdMatchID,omitempty"`
-	ExecCost              float32   `json:"execCost,omitempty"`
-	ExecComm              float32   `json:"execComm,omitempty"`
+	ExecCost              int       `json:"execCost,omitempty"`
+	ExecComm              int       `json:"execComm,omitempty"`
 	HomeNotional          float64   `json:"homeNotional,omitempty"`
 	ForeignNotional       float64   `json:"foreignNotional,omitempty"`
 	TransactTime          time.Time `json:"transactTime,omitempty"`
