+++ model_instrument.go
@@ -31,21 +31,21 @@ type Instrument struct {
 	CalcInterval                   time.Time `json:"calcInterval,omitempty"`
 	PublishInterval                time.Time `json:"publishInterval,omitempty"`
 	PublishTime                    time.Time `json:"publishTime,omitempty"`
-	MaxOrderQty                    float32   `json:"maxOrderQty,omitempty"`
+	MaxOrderQty                    int       `json:"maxOrderQty,omitempty"`
 	MaxPrice                       float64   `json:"maxPrice,omitempty"`
-	LotSize                        float32   `json:"lotSize,omitempty"`
+	LotSize                        int       `json:"lotSize,omitempty"`
 	TickSize                       float64   `json:"tickSize,omitempty"`
-	Multiplier                     float32   `json:"multiplier,omitempty"`
+	Multiplier                     int       `json:"multiplier,omitempty"`
 	SettlCurrency                  string    `json:"settlCurrency,omitempty"`
-	UnderlyingToPositionMultiplier float32   `json:"underlyingToPositionMultiplier,omitempty"`
-	UnderlyingToSettleMultiplier   float32   `json:"underlyingToSettleMultiplier,omitempty"`
-	QuoteToSettleMultiplier        float32   `json:"quoteToSettleMultiplier,omitempty"`
+	UnderlyingToPositionMultiplier int       `json:"underlyingToPositionMultiplier,omitempty"`
+	UnderlyingToSettleMultiplier   int       `json:"underlyingToSettleMultiplier,omitempty"`
+	QuoteToSettleMultiplier        int       `json:"quoteToSettleMultiplier,omitempty"`
 	IsQuanto                       bool      `json:"isQuanto,omitempty"`
 	IsInverse                      bool      `json:"isInverse,omitempty"`
 	InitMargin                     float64   `json:"initMargin,omitempty"`
 	MaintMargin                    float64   `json:"maintMargin,omitempty"`
-	RiskLimit                      float32   `json:"riskLimit,omitempty"`
-	RiskStep                       float32   `json:"riskStep,omitempty"`
+	RiskLimit                      int       `json:"riskLimit,omitempty"`
+	RiskStep                       int       `json:"riskStep,omitempty"`
 	Limit                          float64   `json:"limit,omitempty"`
 	Capped                         bool      `json:"capped,omitempty"`
 	Taxed                          bool      `json:"taxed,omitempty"`
@@ -71,14 +71,14 @@ type Instrument struct {
 	LimitUpPrice                   float64   `json:"limitUpPrice,omitempty"`
 	BankruptLimitDownPrice         float64   `json:"bankruptLimitDownPrice,omitempty"`
 	BankruptLimitUpPrice           float64   `json:"bankruptLimitUpPrice,omitempty"`
-	PrevTotalVolume                float32   `json:"prevTotalVolume,omitempty"`
-	TotalVolume                    float32   `json:"totalVolume,omitempty"`
-	Volume                         float32   `json:"volume,omitempty"`
-	Volume24h                      float32   `json:"volume24h,omitempty"`
-	PrevTotalTurnover              float32   `json:"prevTotalTurnover,omitempty"`
-	TotalTurnover                  float32   `json:"totalTurnover,omitempty"`
-	Turnover                       float32   `json:"turnover,omitempty"`
-	Turnover24h                    float32   `json:"turnover24h,omitempty"`
+	PrevTotalVolume                int       `json:"prevTotalVolume,omitempty"`
+	TotalVolume                    int       `json:"totalVolume,omitempty"`
+	Volume                         int       `json:"volume,omitempty"`
+	Volume24h                      int       `json:"volume24h,omitempty"`
+	PrevTotalTurnover              int       `json:"prevTotalTurnover,omitempty"`
+	TotalTurnover                  int       `json:"totalTurnover,omitempty"`
+	Turnover                       int       `json:"turnover,omitempty"`
+	Turnover24h                    int       `json:"turnover24h,omitempty"`
 	PrevPrice24h                   float64   `json:"prevPrice24h,omitempty"`
 	Vwap                           float64   `json:"vwap,omitempty"`
 	HighPrice                      float64   `json:"highPrice,omitempty"`
@@ -94,8 +94,8 @@ type Instrument struct {
 	ImpactMidPrice                 float64   `json:"impactMidPrice,omitempty"`
 	ImpactAskPrice                 float64   `json:"impactAskPrice,omitempty"`
 	HasLiquidity                   bool      `json:"hasLiquidity,omitempty"`
-	OpenInterest                   float32   `json:"openInterest,omitempty"`
-	OpenValue                      float32   `json:"openValue,omitempty"`
+	OpenInterest                   int       `json:"openInterest,omitempty"`
+	OpenValue                      int       `json:"openValue,omitempty"`
 	FairMethod                     string    `json:"fairMethod,omitempty"`
 	FairBasisRate                  float64   `json:"fairBasisRate,omitempty"`
 	FairBasis                      float64   `json:"fairBasis,omitempty"`
