+++ model_insurance.go
@@ -8,5 +8,5 @@ import (
 type Insurance struct {
 	Currency      string    `json:"currency"`
 	Timestamp     time.Time `json:"timestamp"`
-	WalletBalance float32   `json:"walletBalance,omitempty"`
+	WalletBalance int       `json:"walletBalance,omitempty"`
 }
