+++ model_notification.go
@@ -6,11 +6,11 @@ import (
 
 // Account Notifications
 type Notification struct {
-	Id                float32   `json:"id,omitempty"`
+	Id                int       `json:"id,omitempty"`
 	Date              time.Time `json:"date"`
 	Title             string    `json:"title"`
 	Body              string    `json:"body"`
-	Ttl               float32   `json:"ttl"`
+	Ttl               int       `json:"ttl"`
 	Type_             string    `json:"type,omitempty"`
 	Closable          bool      `json:"closable,omitempty"`
 	Persist           bool      `json:"persist,omitempty"`
