+++ model_stats_usd.go
@@ -1,10 +1,10 @@
 package bitmexgo
 
 type StatsUsd struct {
-	RootSymbol   string  `json:"rootSymbol"`
-	Currency     string  `json:"currency,omitempty"`
-	Turnover24h  float32 `json:"turnover24h,omitempty"`
-	Turnover30d  float32 `json:"turnover30d,omitempty"`
-	Turnover365d float32 `json:"turnover365d,omitempty"`
-	Turnover     float32 `json:"turnover,omitempty"`
+	RootSymbol   string `json:"rootSymbol"`
+	Currency     string `json:"currency,omitempty"`
+	Turnover24h  int    `json:"turnover24h,omitempty"`
+	Turnover30d  int    `json:"turnover30d,omitempty"`
+	Turnover365d int    `json:"turnover365d,omitempty"`
+	Turnover     int    `json:"turnover,omitempty"`
 }
